Add team scores next to the time left on the HUD
Adds Combine score to the left of the time left
Adds Rebels score to the right of the time left

ConVars:
sv_timeleft_teamscore (Def. 1) - Shows teams score on the HUD
sv_timeleft_color_override (Def. 0) - Overrides custom color settings to update colors based on the current winning team (red for rebels, blue for combine, white for a tie). sv_timeleft_teamscore must be enabled to use this! Otherwise, it resets back to 0.